LayoutTests/http/tests/site-isolation/mediastream/getUserMedia-video-starts.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that getUserMedia for video starts in an OOP iframe.");
jsTestIsAsync = true;

if (window.testRunner) {
    window.testRunner.setUserMediaPermission(true);
}

window.addEventListener("message", function (event) {
    shouldBe("event.data", "'success'");
    finishJSTest();
});
</script>
<iframe allow="camera *" src="http://localhost:8000/site-isolation/resources/getUserMedia-video-starts-frame.html"></iframe>
