LayoutTests/http/tests/site-isolation/mediastream/getUserMedia-video-starts-expected.txt

Tests that getUserMedia for video starts in an OOP iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is 'success'
PASS successfullyParsed is true

TEST COMPLETE
